chore: adjust component names
