+++ frontend/src/components/form-toggle/components/switch-label/styles.js
