+++ frontend/src/components/form-toggle/components/switch/styles.js
