+++ frontend/src/components/logo/styles.js
