+++ frontend/src/components/medical-disclaimer/styles.js
