+++ frontend/src/components/results/styles.js
